refactor(external): split zwift_client.py into facade + 2 sub-modules
Extract SQLite cache operations into client/cache.py and search/scoring
logic into client/search.py. ZwiftWorkoutClient remains as a thin facade
that delegates to the sub-modules. Public API is 100% preserved.

- client/cache.py: init_cache_db, cleanup_expired_cache, save/load workouts,
  mark_workout_used, get_cache_stats, search_catalog_db
- client/search.py: get_categories_for_session_type, calculate_match_score,
  is_recently_used
- zwift_client.py: ~120L facade class (was 542L)
