+++ magma_cycling/external/client/__init__.py
@@ -1,0 +1,1 @@
+"""Zwift workout client sub-modules: cache operations and search logic."""
